requirements.txt
paramiko==3.4.0
PyYAML==6.0.1
matplotlib==3.8.2
numpy==1.26.3
